make ServerRequestFactory a final class
