+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - `Innmind\Http\Factory\FilesFactory` is now a final class
 - `Innmind\Http\Factory\FormFactory` is now a final class
 - `Innmind\Http\Factory\QueryFactory` is now a final class
+- `Innmind\Http\Factory\ServerRequestFactory` is now a final class
 
 ### Removed
 
@@ -21,6 +22,7 @@
 - `Innmind\Http\Factory\Files\FilesFactory`
 - `Innmind\Http\Factory\Form\FormFactory`
 - `Innmind\Http\Factory\Query\QueryFactory`
+- `Innmind\Http\Factory\ServerRequest\ServerRequestFactory`
 
 ### Fixed
